Fix missing toolbars at startup on Linux
The problem was that on Linux (not Windows or Mac) menu
events are generated when checking the menu items in
the toolbar menu.

Becuase of the recent changes I made and how the toolbars
are being shown (toggled) when the toolbar menu is checked,
they were simply being hidden.  Actually, the were being
shown and then hidden because of the toggle.

So, with this change I made exposing toolbars a bit more
deterministic.  The menus still use a toggle, but the 
ToolManager now shows or hides the bars explicitly during
setup, so the menu toggle issue is resolved.